Completed first full version of Archive Ingestor Backend
Added reference for session information to ArchiveIngestor class so session values can be directly manipulated. Added checkForCancel function to ArchiveIngestor so it can check if it's thread has been interrupted and placed it in key points of class. Reworked error package with new names and changing some error classes to Exceptions that need to be handled by compile time. Added new async configuration class to Spring Boot backend to allow for asynchronous archive parsing. Reworked API in ArchiveIngestorController and added session-specific paths. Replaced the storyInfo attribute in ArchiveIngestor Request with sessionNickname. Added sessionInfo attribute to ArchiveIngestorResponse and a second constructor. Added new ArchiveIngestorSession, ArchiveIngestorSessionInfo, and ArchiveIngestionTaskFuture models to use for managing parsing sessions. Added a new AsyncTasks class to holder functions that need to run in separate threads. Added additional settings to application.properties. Reworked ArchiveIngestorService to decouple the processing of requests and the parsing of archives. Replaced ArchiveIngestor Service's parseStory and parseChapter with startParseStory and startParseChapter respectively, moving the former functions to Async.  Added session purging, getting, and deleting functions to service class. Made minor updates to testing suite to keep in line with changes.