Fixed two bugs after doing fullstack test run
Encased entirety of story metadata HTML in an if statement so that story metatdata components with undefined storyMetadataResultUnits are not displayed. Fixed bug in ArchiveIngestor backend service where, if the outer element for chapter start notes but there is no userstuff element inside, the ingestor does not handle it properly.